backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synapse-study-api"
version = "0.1.0"
description = "FastAPI backend for the Synapse Study app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "dramatiq[redis]>=1.17.1,<2.0",
  "fastapi>=0.115.0,<1.0",
  "pydantic-settings>=2.6.1,<3.0",
  "pypdf>=5.1.0,<6.0",
  "python-dotenv>=1.0.1,<2.0",
  "python-multipart>=0.0.12,<1.0",
  "python-pptx>=1.0.2,<2.0",
  "redis>=5.2.0,<6.0",
  "sqlalchemy>=2.0.36,<3.0",
  "uvicorn[standard]>=0.32.0,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0",
  "pytest>=8.3.3,<9.0",
  "reportlab>=4.2.5,<5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
